requirements.txt
pandas
numpy
torch
transformers
pillow
tqdm
librosa
nibabel
nilearn
matplotlib
torchvision